updated remove bad library
