+++ package.json
@@ -20,7 +20,6 @@
     "react-color": "^2.19.3",
     "react-dom": "^19.0.0",
     "react-icons": "^5.5.0",
-    "react-scroll-parallax": "^3.4.5",
     "tailwindcss": "^4.0.12",
     "three": "^0.176.0"
   },
