+++ src/app/test6/helpers.ts
